+++ src/kernels.h
@@ -3,7 +3,7 @@
 #include "cuda_runtime.h"
 
 namespace deepc {
-    void bias_add_fwd_gpu(const double* a, const double* b, double* out,
+    void bias_add_fwd_gpu(const float* a, const float* b, float* out,
                             int rows, int cols);
 
     void matmul_tiled_fwd_gpu(const double* a, const double* b, double* out, int M, int N, int K);
